+++ Common/src/main/java/fuzs/pickupnotifier/mixin/client/ClientPacketListenerMixin.java
@@ -1,19 +1,25 @@
 package fuzs.pickupnotifier.mixin.client;
 
 import fuzs.pickupnotifier.client.handler.AddEntriesHandler;
+import net.minecraft.client.Minecraft;
 import net.minecraft.client.multiplayer.ClientPacketListener;
 import net.minecraft.network.protocol.game.ClientGamePacketListener;
 import net.minecraft.network.protocol.game.ClientboundTakeItemEntityPacket;
+import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(ClientPacketListener.class)
-public abstract class ClientPacketListenerMixin implements ClientGamePacketListener {
+abstract class ClientPacketListenerMixin implements ClientGamePacketListener {
+    @Shadow
+    @Final
+    private Minecraft minecraft;
 
     @Inject(method = "handleTakeItemEntity", at = @At(value = "INVOKE", target = "Lnet/minecraft/network/protocol/PacketUtils;ensureRunningOnSameThread(Lnet/minecraft/network/protocol/Packet;Lnet/minecraft/network/PacketListener;Lnet/minecraft/util/thread/BlockableEventLoop;)V", shift = At.Shift.AFTER))
-    public void handleTakeItemEntity$invokeEnsureRunningOnSameThread(ClientboundTakeItemEntityPacket packet, CallbackInfo callback) {
-        AddEntriesHandler.onEntityPickup(packet.getItemId(), packet.getPlayerId(), packet.getAmount());
+    public void handleTakeItemEntity(ClientboundTakeItemEntityPacket packet, CallbackInfo callback) {
+        AddEntriesHandler.onEntityPickup(this.minecraft, packet.getItemId(), packet.getPlayerId(), packet.getAmount());
     }
 }